+++ src/main/java/de/fh_zwickau/informatik/sensor/model/devices/types/SwitchMultilevel.java
@@ -23,8 +23,18 @@ public class SwitchMultilevel extends Device {
     }
 
     @Override
+    public String up() {
+        return mCommandHandler.up(this);
+    }
+
+    @Override
     public String off() {
         return mCommandHandler.off(this);
+    }
+
+    @Override
+    public String down() {
+        return mCommandHandler.down(this);
     }
 
     @Override
@@ -38,6 +48,21 @@ public class SwitchMultilevel extends Device {
     }
 
     @Override
+    public String upMax() {
+        return mCommandHandler.upMax(this);
+    }
+
+    @Override
+    public String increase() {
+        return mCommandHandler.increase(this);
+    }
+
+    @Override
+    public String decrease() {
+        return mCommandHandler.decrease(this);
+    }
+
+    @Override
     public String update() {
         return mCommandHandler.update(this);
     }
@@ -47,9 +72,29 @@ public class SwitchMultilevel extends Device {
         return mCommandHandler.exact(this, level);
     }
 
+    @Override
+    public String exactSmooth(String level, Integer duration) {
+        return mCommandHandler.exactSmooth(this, level, duration);
+    }
+
+    @Override
+    public String stop() {
+        return mCommandHandler.stop(this);
+    }
+
+    @Override
+    public String startUp() {
+        return mCommandHandler.startUp(this);
+    }
+
+    @Override
+    public String startDown() {
+        return mCommandHandler.startDown(this);
+    }
+
     /*
      * (non-Javadoc)
-     * 
+     *
      * @see de.fh_zwickau.informatik.sensor.model.devices.Device#toString()
      */
     @Override